services/workflow_orchestrator/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
httpx>=0.24.0
asyncio>=3.4.3
json-schema>=0.1.0
python-dotenv>=1.0.0
smithery
mcp
